this might do it

+++ esphome/components/wordclock2/wordclock.h
@@ -224,10 +224,10 @@ class WordClockLightOutputBase : public wordclock2::WordClock {
 		if (time.is_valid() == false) {
 			ESP_LOGE("loop", "Got invalid time from current_time Time: %i:%i", h, m );
 			for (int i = 0; i < 8; i++) {
-				int random_number = random(1023);
-				int rRed = random(255);
-				int rGreen = random(255);
-				int rBlue = random(255);
+				uint8_t random_number = random(1023);
+				uint8_t rRed = random(255);
+				uint8_t rGreen = random(255);
+				uint8_t rBlue = random(255);
 				if (random_number %2 == 1) {
 					setPixelColor(thinking[i][0], rRed, rGreen, rBlue, scaledBrightness); show();
 					delay(250);
